Feature/auto updater proxy (#1)
feat: Use Electron.Net to send http requests for auto updater

There was a refactoring around httpRequests and util/restApiRequests, they both forward the calls to an instance of the HttpExecutor interface, which can be Node or Electron. Node is used by default and electron is only used in case it's an electron app running the auto updater
