+++ nsis-auto-updater/package.json
@@ -13,6 +13,7 @@
   ],
   "dependencies": {
     "bluebird-lst-c": "^1.0.5",
+    "electron": "^1.4.12",
     "debug": "^2.4.4",
     "fs-extra-p": "^3.0.3",
     "ini": "^1.3.4",
@@ -21,5 +22,8 @@
     "source-map-support": "^0.4.6",
     "tunnel-agent": "^0.4.3"
   },
-  "typings": "./out/electron-auto-updater.d.ts"
+  "typings": "./out/electron-auto-updater.d.ts",
+  "devDependencies": {
+    "@types/electron": "^1.4.29"
+  }
 }
