+++ src/publish/bintray.ts
@@ -1,6 +1,6 @@
 import { bintrayRequest } from "./restApiRequest"
 import { BintrayOptions } from "../options/publishOptions"
-export { HttpError } from "./restApiRequest"
+export { HttpError } from "../util/httpExecutor"
 
 export interface Version {
   readonly name: string
